+++ src/reducers/tutorials/tutorials.ts
@@ -1,6 +1,7 @@
 import {TUTORIAL_UPDATE, TUTORIALS_FIND} from '../../actions/_types';
 import {tutorialUpdate} from './update-tutorial';
 import RootPackage from '../../services/root-package';
+import {searchForTutorials} from './check-tutorials';
 
 export default function tutorialsReducer(tutorials = [],
   action: CR.Action): CR.Tutorial[] {
@@ -9,7 +10,10 @@ export default function tutorialsReducer(tutorials = [],
       tutorialUpdate(action.payload.name);
       /* falls through */
     case TUTORIALS_FIND:
-      return RootPackage.getTutorials();
+          const packageJson = RootPackage.get();
+          return ([]
+            .concat(searchForTutorials(packageJson.dependencies))
+            .concat(searchForTutorials(packageJson.devDependencies)));
     default:
       return tutorials;
   }