+++ CHANGELOG.md
@@ -3,10 +3,14 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.7.0] - in progress
-- move to 'material-ui@0.15-beta.1'
+- update to 'react@15'
+- move to 'material-ui@0.15'
 - much improved setup ui & checks
   - check Node, NPM versions
   - Stepper UI
+- loading tutorials
+  - check for new versions
+  - update tutorial button
 - remove 'lodash' dependency
 - attach hints to toolbar